pyproject.toml
[project]
name = "robot-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "eclipse-zenoh>=1.8.0",
    "nanobind>=2.12.0",
    "pyyaml>=6.0.3",
    "reduct-py>=1.18.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
